wip: extract GamePlatformRegionService, break circular dep (steps 2-3 of game split)
Work in progress -- paused mid-task due to token limit. See PR description for
what has been done and what remains.
