+++ src/functions/GameIgdbSync.ts
@@ -16,6 +16,7 @@ import { safeIgnore } from "../utilities/AsyncUtils.js";
 import { logError, logWarn } from "../utilities/LogUtils.js";
 import { clearAutocompleteSearchCaches } from "./GameAutocompleteCache.js";
 import Game from "../classes/Game.js";
+import GamePlatformRegionService from "../classes/GamePlatformRegionService.js";
 
 type IGDBReleaseDate = NonNullable<IGDBGameDetails["release_dates"]>[number];
 
@@ -101,7 +102,7 @@ export async function saveReleaseDates(
   for (const { release, date } of earliestByPlatform.values()) {
     if (!release.platform?.id) continue;
 
-    const platform = await Game.ensurePlatform({
+    const platform = await GamePlatformRegionService.ensurePlatform({
       id: release.platform.id,
       name: release.platform.name ?? null,
     });
@@ -111,7 +112,7 @@ export async function saveReleaseDates(
       continue;
     }
 
-    const region = await Game.ensureRegion(release.region ?? 8);
+    const region = await GamePlatformRegionService.ensureRegion(release.region ?? 8);
     if (!region) continue;
 
     const format: "Physical" | "Digital" | null = null;
@@ -120,6 +121,35 @@ export async function saveReleaseDates(
   }
 
   await updateInitialReleaseDate(gameId);
+}
+
+export async function clearReleaseDates(
+  gameId: number,
+): Promise<{ releases: number; announcements: number }> {
+  return dbTransaction(async (conn) => {
+    const announceCount = await dbMutateConn(
+      conn,
+      GameSql.clearReleaseAnnouncements,
+      { gameId },
+    );
+    const releaseCount = await dbMutateConn(conn, GameSql.clearReleases, {
+      gameId,
+    });
+    await dbMutateConn(conn, GameSql.clearInitialReleaseDate, { gameId });
+    return {
+      releases: Number(releaseCount),
+      announcements: Number(announceCount),
+    };
+  });
+}
+
+export async function refreshReleaseDates(
+  gameId: number,
+  releases: NonNullable<IGDBGameDetails["release_dates"]>,
+): Promise<void> {
+  await clearReleaseDates(gameId);
+  if (!releases.length) return;
+  await saveReleaseDates(gameId, releases);
 }
 
 export async function saveFullGameMetadata(
@@ -336,7 +366,7 @@ export async function addGamePlatformsByIgdbIds(
   );
   if (!uniqueIds.length) return;
 
-  const platformMap = await Game.getPlatformsByIgdbIds(uniqueIds);
+  const platformMap = await GamePlatformRegionService.getPlatformsByIgdbIds(uniqueIds);
   const missingIds = uniqueIds.filter((id) => !platformMap.has(id));
   if (missingIds.length) {
     logWarn(
